+++ src/CP Interpreter/token.cpp
@@ -6,5 +6,5 @@ using namespace lexer;
 
 Token::Token() = default;
 
-Token::Token(TOKEN_TYPE type, std::string value, unsigned long long row, unsigned long long col)
+Token::Token(TOKEN_TYPE type, std::string value, unsigned int row, unsigned int col)
 	: type(type), value(value), row(row), col(col) {}
